c-practice

+++ c-practice/q3/revnum.c
@@ -1,15 +1,15 @@
 #include<stdio.h>
 int main(){
-    int n,r,d;
+    long long int n,r,d;
     r=0;
     printf("enter the number to reverse\n");
-    scanf("%d",&n);
+    scanf("%lld",&n);
     while(n>0)
     {
         d=n%10;
         r=(r*10)+d;
         n=n/10;
     }
-    printf("%d\n",r);
+    printf("%lld\n",r);
     return 0;
 }
